Hadoop: Create a wrapper for exploit implementation
